+++ src/views/SadBandNames/SadBandNames.vue
@@ -49,6 +49,10 @@
         <p class="font-bold">
           How rock it is: <span class="font-extralight">{{ band.score }}%</span>
         </p>
+        <p class="font-bold">
+          User Opinion:
+          <span class="font-extralight">{{ band.comment }}</span>
+        </p>
       </el-card>
 
       <p v-show="savedBandNames.length <= 0" class="text-gray-600 mt-3">
@@ -68,19 +72,6 @@ import CustomCard from "@/components/common/CustomCard.vue";
 const store = useStore();
 const bandName = computed(() => store.state.bandNames.bandName);
 const savedBandNames = computed(() => store.state.bandNames.savedBandNames);
-
-const comments = [
-  "Your band name is on fire!",
-  "Rock'n'roll all night with this band name!",
-  "The crowd will be chanting this name for sure.",
-  "This name has a legendary vibe.",
-  "Get ready to hit the stage with this name!",
-];
-
-const generateRandomComment = () => {
-  const randomIndex = Math.floor(Math.random() * comments.length);
-  return comments[randomIndex];
-};
 
 const generateName = () => {
   store.commit("generateBandName");
